use xpcconnectionhandle instead of IntPtr
