+++ DragonFruit.OnionFruit.Core.MacOS/Native/NativeMethods.cs
@@ -3,6 +3,17 @@ using System.Runtime.InteropServices;
 
 namespace DragonFruit.OnionFruit.Core.MacOS.Native;
 
+internal enum NativeStatus
+{
+    Ok = 0,
+    NetworkServiceNotFound = 1,
+    NetworkProtocolNotSupported = 2,
+    ConfigurationUpdateFailed = 3,
+    XpcConnectionFailed = 4,
+    XpcRequestTimeout = 5,
+    XpcVersionMismatch = 6,
+}
+
 internal static partial class NativeMethods
 {
     private const string LibraryName = "libonionfruit";
@@ -14,30 +25,33 @@ internal static partial class NativeMethods
     public static partial void DestroyNetworkServiceList(IntPtr list, int count);
 
     [LibraryImport(LibraryName, EntryPoint = "createXpcConnection", StringMarshalling = StringMarshalling.Utf8)]
-    public static partial NativeStatus CreateXpcConnection(string serviceName, out IntPtr connectionRef, out int serverVersion);
+    public static partial NativeStatus CreateXpcConnection(string serviceName, out XpcConnectionHandle connectionRef, out int serverVersion);
 
     [LibraryImport(LibraryName, EntryPoint = "destroyXpcConnection")]
-    public static partial void DestroyXpcConnection(IntPtr connection);
+    public static partial void DestroyXpcConnection(XpcConnectionHandle connection);
 
     [LibraryImport(LibraryName, EntryPoint = "getServiceProxyConfig", StringMarshalling = StringMarshalling.Utf8)]
-    public static partial NativeStatus GetServiceProxyConfig(IntPtr xpcConnection, string serviceId, out IntPtr proxyConfigPtr);
+    public static partial NativeStatus GetServiceProxyConfig(XpcConnectionHandle xpcConnection, string serviceId, out IntPtr proxyConfigPtr);
 
     [LibraryImport(LibraryName, EntryPoint = "destroyServiceProxyConfig")]
     public static partial void DestroyProxyConfig(IntPtr proxyConfigPtr);
 
     [LibraryImport(LibraryName, EntryPoint = "getServiceDnsResolvers", StringMarshalling = StringMarshalling.Utf8)]
-    public static partial NativeStatus GetServiceDnsResolvers(IntPtr xpcConnection, string serviceId, out IntPtr resolverListPtr, out int resolverCount);
+    public static partial NativeStatus GetServiceDnsResolvers(XpcConnectionHandle xpcConnection, string serviceId, out IntPtr resolverListPtr, out int resolverCount);
 
     [LibraryImport(LibraryName, EntryPoint = "destroyServiceDnsResolvers")]
     public static partial void DestroyDnsResolverList(IntPtr resolverListPtr);
 
     [DllImport(LibraryName, EntryPoint = "setServiceDnsResolvers", ExactSpelling = true)]
     public static extern NativeStatus SetServiceDnsResolvers(
-        IntPtr xpcConnection,
+        XpcConnectionHandle xpcConnection,
         [In, MarshalAs(UnmanagedType.LPUTF8Str)] string serviceId,
         [In, MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.LPUTF8Str)] string[] resolvers,
         int resolverCount);
 
     [DllImport(LibraryName, EntryPoint = "setServiceProxyConfig", ExactSpelling = true)]
-    public static extern NativeStatus SetServiceProxyConfig(IntPtr xpcConnection, [In, MarshalAs(UnmanagedType.LPUTF8Str)] string serviceId, [In] ServiceProxyConfig configPtr);
+    public static extern NativeStatus SetServiceProxyConfig(XpcConnectionHandle xpcConnection, [In, MarshalAs(UnmanagedType.LPUTF8Str)] string serviceId, [In] ServiceProxyConfig configPtr);
+
+    [LibraryImport(LibraryName, EntryPoint = "showMessageBox", StringMarshalling = StringMarshalling.Utf8)]
+    public static partial void ShowMessageBox(string title, string message, string buttonText);
 }
